+++ Project_KFC_WEB/Areas/Admin/Controllers/FoodCategoriesController.cs
@@ -18,6 +18,9 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/FoodCategories
         public ActionResult Index(int page = 1, bool isReset = false)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
 
             if (isReset) Session["isSearchingFoodCategory"] = false;
 
@@ -80,6 +83,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/FoodCategories/Details/5
         public ActionResult Details(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -95,6 +102,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/FoodCategories/Create
         public ActionResult Create()
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             return View();
         }
 
@@ -105,6 +116,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "id,name,image")] foodCategory foodCategory, HttpPostedFileBase imageFile)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (ModelState.IsValid)
             {
                 if (imageFile != null && imageFile.ContentLength > 0)
@@ -137,6 +152,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/FoodCategories/Edit/5
         public ActionResult Edit(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -156,6 +175,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id,name,image")] foodCategory foodCategory, HttpPostedFileBase imageFile)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (ModelState.IsValid)
             {
                 if (imageFile != null && imageFile.ContentLength > 0)
@@ -193,6 +216,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/FoodCategories/Delete/5
         public ActionResult Delete(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -210,6 +237,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             foodCategory foodCategory = db.foodCategories.FirstOrDefault((item) => item.id == id);
             db.foodCategories.Remove(foodCategory);
             db.SaveChanges();
