+++ Project_KFC_WEB/Areas/Admin/Controllers/FoodsController.cs
@@ -18,6 +18,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/Foods
         public ActionResult Index(int page = 1, bool isReset = false)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (isReset) Session["isSearchingFood"] = false;
 
             List<food> foods = new List<food>();
@@ -124,6 +128,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/Foods/Details/5
         public ActionResult Details(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -150,6 +158,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "id,idCategory,name,image,price,discount,description,timeSellStart,timeSellEnd")] food food, HttpPostedFileBase imageFile)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (ModelState.IsValid)
             {
                 if (imageFile != null && imageFile.ContentLength > 0)
@@ -182,6 +194,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/Foods/Edit/5
         public ActionResult Edit(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -202,6 +218,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Edit([Bind(Include = "id,idCategory,name,image,price,discount,description,timeSellStart,timeSellEnd")] food food, HttpPostedFileBase imageFile)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (ModelState.IsValid)
             {
                 if (imageFile != null && imageFile.ContentLength > 0)
@@ -245,6 +265,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         // GET: Admin/Foods/Delete/5
         public ActionResult Delete(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
@@ -262,6 +286,10 @@ namespace Project_KFC_WEB.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int? id)
         {
+            bool isLogin = Session["login"] != null ? (bool)Session["login"] : false;
+
+            if (!isLogin) return RedirectToAction("Login", "Home", new { isLogin = false });
+
             food food = db.foods.Find(id);
             db.foods.Remove(food);
             db.SaveChanges();
